changed accounts.cpp to where I can get it to work. Fix required moving cpp code back, and moving includes to .h file
